[CI] Add Github Actions and fix bug (#2)
* feat: add github actions bout macos and linux

* fix: consider installed or not

* fix: add brew config to .zprofile

* chore: change shell to zsh

* fix: change `brew doctor` order backward

* chore: add check commands in GHA

* refactor: put linux and mac workflows together

* chore: remove unnecessary libraries

* fix: enable to install rustup with no-prompt

* chore: change set options

* chore: install brew with non-interactive

* chore: update workflow

* fix: exec zsh as login shell

* chore: run workflow when labeled

+++ scripts/mac/defaults.sh
@@ -1,4 +1,6 @@
-#!/bin/bash -eu
+#!/bin/bash
+
+set -eux
 
 if [ "$(uname)" != "Darwin" ]; then
 	exit 0
@@ -49,7 +51,9 @@ killall Finder
 # VSCode
 #
 
-# for nvim
-defaults write com.microsoft.VSCode ApplePressAndHoldEnabled -bool false
+if builtin command -v code; then
+	# for nvim
+	defaults write com.microsoft.VSCode ApplePressAndHoldEnabled -bool false
 
-killall VSCode
+	killall VSCode
+fi
